Split large minesweeper_solver.ry for readability
